% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_ci}
\alias{calculate_ci}
\alias{calculate}
\title{Calculate PsN bootstrap confidence intervals}
\usage{
calculate_ci(x, parameter = "THETA3", ci_width = 95, exclusions = c(1, 2))
}
\arguments{
\item{x}{A tibble corresponding to PsN bootstrap output. \code{calculate_ci()} is meant to be run after running \code{combine_bootstraps()} but will also work with an imported \code{raw_results_run#.csv} file created by PsN.}

\item{parameter}{The parameter of interest. Defaults to \code{THETA3}.}

\item{ci_width}{The desired confidence interval width. Defaults to \code{95} for a 95\code{\%} confidence interval.}

\item{exclusions}{A numeric vector corresponding to the desired PsN exclusions. Defaults to \code{c(1, 2)} for exclusions 1 and 2, the default PsN exclusions. The available exclusions are:
\itemize{
\item Exclusion 1: Skip if \code{minimization_successful == 0}.
\item Exclusion 2: Skip if \code{estimate_near_boundary == 1}.
\item Exclusion 3: Skip if \code{covariance_step_successful == 0}.
\item Exclusion 4: Skip if \code{covariance_step_warnings == 1}.
}}
}
\value{
A tibble that contains:
\itemize{
\item \code{n_original}: The total number of rows.
\item \code{n_used}: The number of successful runs used to calculate the confidence interval.
\item \code{n_excluded}: The number of excluded runs.
\item \code{parameter}: The parameter of interest.
\item \code{ci_width}: The width of the confidence interval.
\item \code{lower_ci}: The lower value of the calculated confidence interval.
\item \code{point_estimate}: The point estimate from the original dataset.
\item \code{upper_ci}: The upper value of the calculated confidence interval.
}

\code{calculate_ci()} also provides verbose output in the console listing the values of \code{n_original}, \code{n_used}, \code{n_excluded}, the exclusions used, and a printed tibble with the variables above.
}
\description{
\code{calculate_ci()} calculates confidence intervals from PsN bootstraps.
}
