% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_bootstraps}
\alias{combine_bootstraps}
\alias{combine}
\title{Combine PsN bootstraps}
\usage{
combine_bootstraps(path, runs = "bs_run1", out = "bootstrap_combined.csv")
}
\arguments{
\item{path}{The path to the working directory containing PsN bootstrap folders. Defaults to \code{here::here()}.}

\item{runs}{A character vector of folder names to be searched. Defaults to \code{bs_run1} per default PsN settings.}

\item{out}{The name of the output .csv file. Defaults to \code{bootstrap_combined.csv}.}
}
\value{
A tibble containing the row-bound model estimates of PsN bootstraps.
}
\description{
\code{combine_bootstraps()} extracts and row-binds model estimates of PsN bootstraps into a tibble.
}
\details{
\code{combine_bootstraps()} searches inside all folders specified by the \code{runs} argument for model estimates of PsN bootstraps.

By PsN convention, each bootstrap folder has a file of the general form \code{raw_results_run#.csv} that contains estimates for the original and bootstrapped models.
The first row in each \code{raw_results_run#.csv} file corresponds to the original dataset.
To avoid duplicating the row corresponding to the original dataset, \code{combine_bootstraps()} keeps only the first row of the first .csv file passed to the \code{runs} argument.
\code{combine_bootstraps()} throws an informative warning if the column names or the first row of each \code{raw_results_run#.csv} file are not identical.

\code{combine_bootstraps()} creates a new .csv file in the working directory, specified by the \code{out} argument, that contains the combined estimates of PsN bootstraps.
}
