% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizons.R
\name{horizons_get_ephemeris}
\alias{horizons_get_ephemeris}
\title{Retrieve and parse Horizons ephemeris}
\usage{
horizons_get_ephemeris(...)
}
\value{
data.frame with parsed Horizons ephemeris and POSIXct time
}
\description{
Retrieve and parse Horizons ephemeris
}
