% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizons.R
\name{horizons_query}
\alias{horizons_query}
\title{Query NASA JPL Horizons API (text output)}
\usage{
horizons_query(
  target = "301",
  lon,
  lat,
  elev = 0,
  start_time,
  stop_time,
  step = "1d",
  quantities = c("4", "8", "9"),
  refracted = TRUE
)
}
\arguments{
\item{target}{Horizons target ID (e.g. 301 for the Moon, 10 for the Sun)}

\item{lon}{Longitude in decimal degrees (east positive)}

\item{lat}{Latitude in decimal degrees (north positive)}

\item{elev}{Altitude above ITRF93 (or WGS-84 GPS) ellipsoid (m)}

\item{start_time}{Start time, as POSIXct with the local time zone}

\item{stop_time}{Stop time, as POSIXct}

\item{step}{Step size (e.g. "1h", "1d")}

\item{quantities}{Character vector of Horizons quantity (variables) codes (https://ssd.jpl.nasa.gov/horizons/manual.html#obsquan). Default to `c("4", "8", "9")`.}

\item{refracted}{Logical, whether to request refracted apparent coordinates of the target, default to TRUE}
}
\value{
Character string containing the Horizons text output
}
\description{
Retrieve ephemeris for a given target (Sun, Moon...) from a site on Earth using the
Horizons API, returning the raw text output.
Details about parameters: https://ssd-api.jpl.nasa.gov/doc/horizons.html
}
