% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizons.R
\name{illuminance_from_horizons}
\alias{illuminance_from_horizons}
\title{Convert Horizons magnitude to direct illuminance}
\usage{
illuminance_from_horizons(APmag, mag_ex = 0)
}
\arguments{
\item{APmag}{Apparent visual magnitude from Horizons}

\item{mag_ex}{Atmospheric extinction (magnitudes)}
}
\value{
Illuminance in lux (NA when the target is above the horizon)
}
\description{
Convert Horizons magnitude to direct illuminance
}
