% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtools-package.R
\docType{package}
\name{nmtools-package}
\alias{nmtools}
\alias{nmtools-package}
\title{nmtools: Tools for NONMEM and Perl-speaks-NONMEM (PsN)}
\description{
Tools for common NONMEM and Perl-speaks-NONMEM
    (PsN) workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gvelasq/nmtools}
  \item Report bugs at \url{https://github.com/gvelasq/nmtools/issues}
}

}
\author{
\strong{Maintainer}: Gustavo Velásquez \email{gvelasq2github@gmail.com}

}
\keyword{internal}
