% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_metar.R
\name{interpolate_metar}
\alias{interpolate_metar}
\title{Interpolate Missing Hourly METAR Data}
\usage{
interpolate_metar(df, interp_cols, time_interval = "1 hour")
}
\arguments{
\item{df}{A `data.frame` containing METAR data. It must include a column named `METAR_Date` (POSIXct or coercible)
and numeric columns to interpolate (e.g., Wind_speed, Temperature).}

\item{interp_cols}{A character vector specifying the names of numeric columns to interpolate.}

\item{time_interval}{A character string specifying the time step for interpolation (default is `"1 hour"`).
Passed to `seq.POSIXt(by = time_interval)`.}
}
\value{
A `data.frame` with missing timestamps filled and numeric variables interpolated.
        New rows added by interpolation are flagged with `"This data were interpolated"` in the `Remark` column.
}
\description{
This function fills missing time-stamped rows in a METAR dataset and linearly interpolates numeric weather variables
(e.g., wind speed, temperature, pressure) based on adjacent observations. It is useful for reconstructing hourly
time series when some observations are missing.
}
\examples{
\dontrun{
Metar <- metar_get_historical(
airport = "DFOO",
start_date = "2018-11-01",
end_date = "2018-11-30",
from = "ogimet") \%>\%
metar_decode()
interpolated_data <- interpolate_metar(Metar)
# Check interpolated rows
interpolated_data \%>\% filter(Remark == "These data were interpolated")
}
}
